Better adhere to the Single Responsibility Principle by splitting the WACZ/WARC utilities
