+++ scrapy_webarchive/downloadermiddlewares.py
@@ -7,16 +7,14 @@ from scrapy.crawler import Crawler
 from scrapy.exceptions import IgnoreRequest, NotConfigured
 from scrapy.http.request import Request
 from scrapy.http.response import Response
-from scrapy.responsetypes import ResponseTypes
 from scrapy.settings import Settings
 from scrapy.spiders import Spider
 from scrapy.statscollectors import StatsCollector
 from smart_open import open
 from typing_extensions import Self
-from warc.warc import WARCRecord
 
-from scrapy_webarchive.exceptions import WaczMiddlewareException
-from scrapy_webarchive.utils import MultiWACZFile, WACZFile
+from scrapy_webarchive.wacz import MultiWaczFile, WaczFile
+from scrapy_webarchive.warc import record_transformer
 
 
 class WaczMiddleware:
@@ -26,8 +24,6 @@ class WaczMiddleware:
     Loads the index fully into memory, but lazily loads pages.
     This helps to work with large archives, including remote ones.
     """
-
-    response_types = ResponseTypes()
 
     def __init__(self, settings: Settings, stats: StatsCollector) -> None:
         self.stats = stats
@@ -52,10 +48,10 @@ class WaczMiddleware:
 
         if len(self.wacz_urls) == 1:
             spider.logger.info(f"[WACZDownloader] Opening WACZ {self.wacz_urls[0]}")
-            self.wacz = WACZFile(open(self.wacz_urls[0], "rb", transport_params=tp))
+            self.wacz = WaczFile(open(self.wacz_urls[0], "rb", transport_params=tp))
         else:
             spider.logger.info(f"[WACZDownloader] Opening WACZs {self.wacz_urls}")
-            self.wacz = MultiWACZFile(
+            self.wacz = MultiWaczFile(
                 [open(u, "rb", transport_params=tp) for u in self.wacz_urls]
             )
 
@@ -81,7 +77,7 @@ class WaczMiddleware:
 
                 # do not filter to allow all occurences to be handled
                 # since we don't yet get all information for the request, this can be necessary
-                yield self._request_for_record(
+                yield record_transformer.request_for_record(
                     entry,
                     flags=["wacz_start_request"],
                     meta={"wacz_index_entry": entry},
@@ -104,7 +100,7 @@ class WaczMiddleware:
         # get record from existing index entry, or else lookup by URL
         record = self.wacz.get_record(request.meta.get("wacz_index_entry", request.url))
         if record:
-            response = self._response_for_record(record)
+            response = record_transformer.response_for_record(record)
 
             if not response:
                 self.stats.inc_value("wacz/response_not_recognized", spider=spider)
@@ -116,59 +112,3 @@ class WaczMiddleware:
             # when page not found in archive, return 404, and record it in a statistic
             self.stats.inc_value("wacz/response_not_found", spider=spider)
             return Response(url=request.url, status=404)
-
-    def _request_for_record(self, record: WARCRecord, **kwargs):
-        # TODO locate request in WACZ and include all relevant things (like headers)
-        return Request(url=record["url"], method=record.get("method", "GET"), **kwargs)
-
-    def _response_for_record(self, record: WARCRecord, **kwargs):
-        # We expect a response.
-        # https://iipc.github.io/warc-specifications/specifications/warc-format/warc-1.1/#warc-type-mandatory
-        if record["WARC-Type"] != "response":
-            raise WaczMiddlewareException(f"Unexpected record type: {record['type']}")
-
-        # We only know how to handle application/http.
-        # https://iipc.github.io/warc-specifications/specifications/warc-format/warc-1.1/#content-type
-        record_content_type = (record["Content-Type"] or "").split(";", 1)[0]
-        if record_content_type != "application/http":
-            raise WaczMiddlewareException(
-                f"Unexpected record content-type: {record_content_type}"
-            )
-
-        # There is a date field in record['WARC-Date'], but don't have a use for it now.
-        # https://iipc.github.io/warc-specifications/specifications/warc-format/warc-1.1/#warc-date-mandatory
-
-        payload = record.payload.read()
-        payload_parts = payload.split(b"\r\n\r\n", 1)
-        header_lines = payload_parts[0] if len(payload_parts) > 0 else ""
-        body = payload_parts[1] if len(payload_parts) > 1 else None
-
-        header_lines = header_lines.split(b"\r\n")
-        header_parts = header_lines[0].split(None, 2)
-        protocol = header_parts[0] if len(header_parts) > 0 else None
-        status = header_parts[1] if len(header_parts) > 1 else None
-        headers = header_lines_to_dict(header_lines[1:])
-
-        if not status or not protocol:
-            return None
-
-        response_cls = self.response_types.from_headers(headers)
-
-        return response_cls(
-            url=record["WARC-Target-URI"],
-            status=int(status.decode()),
-            protocol=protocol.decode(),
-            headers=headers,
-            body=body,
-            **kwargs,
-        )
-
-
-def header_lines_to_dict(lines):
-    # XXX only supports each header appearing once, not multiple occurences
-    headers = {}
-    for line in lines:
-        k, v = line.split(b":", 1)
-        v = v.lstrip()
-        headers[k] = v
-    return headers
